feat: algo base funcs
